fix: restore green build under pinned Rust 1.92 (fmt + clippy + advisory)
The autonomous P0-P3 swarm landed all 99 issues but pushed fmt/clippy/advisory
debt that fails the pinned-1.92 CI gate. Behavior-preserving cleanup so main
builds and CI is green:
- rustfmt across the workspace.
- clippy: low-byte via to_le_bytes; checked/widening casts instead of truncating
  `as`; removed needless derefs/conversions/vec!; while-let and single-serve
  if in network loops; assert! for literal-bool/const asserts; justified allows
  for inherent-shape lints; safety-comment adjacency; collapsed a duplicated
  #[test] and restored an orphaned round-trip test.
- deny: ignore RUSTSEC-2025-0134 (rustls-pemfile unmaintained; moved to pki-types).

+++ crates/network/src/reconnect.rs
@@ -84,7 +84,7 @@ impl ReconnectBackoff {
     /// then the base is grown for the next attempt.
     pub fn next_delay<R: FnMut() -> u64>(&mut self, mut rng_u64: R) -> Duration {
         let base = self.current;
-        let base_ns = u128::from(base.as_nanos().min(u128::from(u64::MAX)));
+        let base_ns = base.as_nanos().min(u128::from(u64::MAX));
         let delay = if base_ns == 0 {
             Duration::ZERO
         } else {
